+++ src/main/kotlin/com/example/Employee/controllers/EmployeeController.kt
@@ -19,7 +19,7 @@ import java.util.*
 
 @RestController
 @RequestMapping("/employees")
-public class EmployeeController (private val employeeService: EmployeeService, private val bankAccountservice: BankAccountService){
+public class EmployeeController (private val employeeService: EmployeeService){
 
 
     @PostMapping
@@ -49,35 +49,6 @@ public class EmployeeController (private val employeeService: EmployeeService, p
         employeeService.updateEmployee(employeeId,employee)
         return ResponseEntity.ok((employee))
     }
-
-   @PostMapping("/{employeeId}/bank_accounts")
-    fun addAccount(@PathVariable employeeId: Int, @RequestBody bankAccountrequest: BankAccountRequest): ResponseEntity<BankAccountRequest>{
-       bankAccountservice.addAccount(employeeId,bankAccountrequest)
-       return ResponseEntity.ok(bankAccountrequest)
-    }
-
-  @GetMapping("/{employeeId}/bank_accounts")
-   fun getAccounts(@PathVariable employeeId: Int): ResponseEntity<List<BankAccount>> {
-       return ResponseEntity.ok(bankAccountservice.getAccounts(employeeId))
-   }
-
-
-    @GetMapping("/{employeeId}/bank_accounts/{id}")
-    fun getAccount(@PathVariable employeeId: Int, @PathVariable id: Int): ResponseEntity<BankAccount> {
-        return ResponseEntity.ok(bankAccountservice.getAccount(employeeId,id))
-    }
-
-    @PutMapping("/{employeeId}/bank_accounts/{id}")
-    fun updateAccount(@PathVariable employeeId: Int, @PathVariable id: Int, @RequestBody bankAccountrequest: BankAccountRequest): ResponseEntity<BankAccountRequest>{
-        bankAccountservice.updateAccount(employeeId,id,bankAccountrequest)
-        return ResponseEntity.ok(bankAccountrequest)
-    }
-
-    @DeleteMapping("/{employeeId}/bank_accounts/{id}")
-    fun deleteAccount(@PathVariable employeeId: Int, @PathVariable id: Int){
-        bankAccountservice.deleteAccount(employeeId,id)
-    }
-
 
 }
